+++ ReleaseNotes.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Support for the BMP388 temperature sensor
+
 ## [1.1.0] - 2022-03-13
 
 ### Added